requirements.txt
aniso8601==9.0.1
click==8.0.1
colorama==0.4.4
coverage==5.5
Faker==14.0.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Werkzeug==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
SQLAlchemy==1.4.15
celery==5.2.2
moviepy==1.0.3
numpy==1.22.0
imageio==2.14.0
imageio-ffmpeg==0.4.9
decorator==4.4.2
proglog==0.1.9
psycopg2
tqdm==4.62.0
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
gunicorn==20.1.0
tzdata==2023.3
Redis
google-cloud-storage
google-cloud-pubsub